Working cracking routine (given a guessed key length)
Had to relax the constraint on the keyspace definition, but now
results in viable components of a key.

The project ciphertext is also easliy cracked now after using a very
basic algorithm of looking for viable printable characters and then
for the tokens "and" and "the" with spaces around them.